Feat: 세션 만료 시 로그인 화면으로 이동하는 기능 추가

+++ src/main/java/com/wodnj5/board/controller/CommentController.java
@@ -26,7 +26,10 @@ public class CommentController {
     }
 
     @GetMapping("/comment/delete/{postId}/{commentId}")
-    public String delete(@PathVariable Long postId, @PathVariable Long commentId ) {
+    public String delete(@PathVariable Long postId, @PathVariable Long commentId, @SessionAttribute(name = "user", required = false) User user ) {
+        if(user == null) {
+            return "redirect:/user/login";
+        }
         commentService.delete(commentId);
         return "redirect:/post/read/" + postId;
     }
